Remove unneeded const.
Having const twice in a member function is redundant. In the case
where a copy of a member variable is returned, the const is just
plain mean. For example:
const uint foo() const;

The const on the end is more then enough to allow the compiler to
optimize the member function. The const at the beginning is saying,
"Here, have a copy of something, but you better not change it."