Also use Qt version suffix for CMAKECONFIG_INSTALL_DIR
Otherwise cmake can't find it. Apparently the name of the config
file has to case-insensitively match the name of the dir it's
installed into. While the documentation [1] contains "<name>*" when
listing possible locations, indicating that globbing might happen, it
states a bit below "In all cases the <name> is treated as
case-insensitive and *corresponds* to any of the names specified
(<package> or names given by NAMES)".

[1] https://cmake.org/cmake/help/v3.10/command/find_package.html