Fix parsing of timestamp for EpisodeAction
I don't know the entire history of the API, but the current timestamp
that is returned by gpodder.net is in ISO 8601 format.  The libmygpo-qt
implementation was trying to convert this string directly into
qulonglong which would fail, and therefore always return 0.
Presumably, the timestamp was previously returned as seconds-since-epoch,
but changed at some point in time.